tidy up main
