+++ src/warnet/control.py
@@ -4,7 +4,9 @@ import tempfile
 import time
 
 import click
+import inquirer
 import yaml
+from inquirer.themes import GreenPassion
 from rich import print
 from rich.console import Console
 from rich.prompt import Confirm, Prompt
@@ -248,3 +250,50 @@ def run(scenario_file: str, additional_args: tuple[str]):
         print(f"Failed to start scenario: {scenario_name}")
 
     os.unlink(temp_file_path)
+
+
+@click.command()
+@click.argument("pod_name", type=str, default="")
+@click.option("--follow", "-f", is_flag=True, default=False, help="Follow logs")
+def logs(pod_name: str, follow: bool):
+    """Show the logs of a pod"""
+    follow_flag = "--follow" if follow else ""
+    namespace = get_default_namespace()
+
+    if pod_name:
+        try:
+            command = f"kubectl logs pod/{pod_name} -n {namespace} {follow_flag}"
+            stream_command(command)
+            return
+        except Exception as e:
+            print(f"Could not find the pod {pod_name}: {e}")
+
+    try:
+        pods = run_command(f"kubectl get pods -n {namespace} -o json")
+        pods = json.loads(pods)
+        pod_list = [item["metadata"]["name"] for item in pods["items"]]
+    except Exception as e:
+        print(f"Could not fetch any pods in namespace {namespace}: {e}")
+        return
+
+    if not pod_list:
+        print(f"Could not fetch any pods in namespace {namespace}")
+        return
+
+    q = [
+        inquirer.List(
+            name="pod",
+            message="Please choose a pod",
+            choices=pod_list,
+        )
+    ]
+    selected = inquirer.prompt(q, theme=GreenPassion())
+    if selected:
+        pod_name = selected["pod"]
+        try:
+            command = f"kubectl logs pod/{pod_name} -n {namespace} {follow_flag}"
+            stream_command(command)
+        except Exception as e:
+            print(f"Please consider waiting for the pod to become available. Encountered: {e}")
+    else:
+        pass  # cancelled by user
